kairumi_inokaze/requirements.txt
python-telegram-bot==20.7
flask
openai==1.3.0
apscheduler
python-dotenv
aiohttp
replit
